tools/requirements.txt
# Requirements for quantum readiness checker
py-cpuinfo>=9.0.0
psutil>=5.9.0
numpy>=1.24.0